+++ gateway/src/apicast/policy_chain.lua
@@ -60,7 +60,8 @@ end
 local DEFAULT_POLICIES = {
     'apicast.policy.load_configuration',
     'apicast.policy.find_service',
-    'apicast.policy.local_chain'
+    'apicast.policy.local_chain',
+    'apicast.policy.nginx_metrics'
 }
 
 --- Return new policy chain with default policies.